move legacy text input
