+++ src/screens/AddCash/AddCash.js
@@ -38,7 +38,7 @@ import { isLogV2AppEvents } from 'utils/environment';
 // Components
 import { Container } from 'components/modern/Layout';
 import Button from 'components/modern/Button';
-import TextInput from 'components/TextInput';
+import TextInput from 'components/legacy/TextInput';
 import HeaderBlock from 'components/HeaderBlock';
 import Text from 'components/modern/Text';
 import Toast from 'components/Toast';
