+++ src/screens/ImportWallet/ImportWallet.js
@@ -31,7 +31,7 @@ import { resetWalletImportErrorAction, importWalletFromMnemonicAction } from 'ac
 import { ScrollWrapper, Wrapper } from 'components/Layout';
 import ContainerWithHeader from 'components/Layout/ContainerWithHeader';
 import { BaseText, MediumText } from 'components/Typography';
-import TextInput from 'components/TextInput';
+import TextInput from 'components/legacy/TextInput';
 import Tabs from 'components/Tabs';
 import Button from 'components/Button';
