+++ storybook/storyLoader.js
@@ -30,7 +30,7 @@ function loadStories() {
 	require('../src/components/SWActivationCard/SWActivationCard.stories');
 	require('../src/components/legacy/SearchBar/SearchBar.stories');
 	require('../src/components/Table/Table.stories');
-	require('../src/components/TextInput/TextInput.stories');
+	require('../src/components/legacy/TextInput/TextInput.stories');
 	require('../src/components/Toast/ToastCard.stories');
 	require('../src/components/Tooltip/Tooltip.stories');
 	require('../src/components/legacy/ValueInput/ValueInput.stories');
@@ -64,7 +64,7 @@ const stories = [
 	'../src/components/SWActivationCard/SWActivationCard.stories',
 	'../src/components/legacy/SearchBar/SearchBar.stories',
 	'../src/components/Table/Table.stories',
-	'../src/components/TextInput/TextInput.stories',
+	'../src/components/legacy/TextInput/TextInput.stories',
 	'../src/components/Toast/ToastCard.stories',
 	'../src/components/Tooltip/Tooltip.stories',
 	'../src/components/legacy/ValueInput/ValueInput.stories',
